feat: finish fix unit test

+++ Application/Features/Suppliers/Commands/DeleteSupplier/DeleteSupplierCommandHandler.cs
@@ -1,7 +1,7 @@
 using Application.Common.Models;
 using Application.Interfaces.Repositories;
 using Application.Interfaces.Repositories.Supplier;
-
+using Domain.Constants.Input;
 using MediatR;
 
 namespace Application.Features.Suppliers.Commands.DeleteSupplier;
@@ -18,6 +18,11 @@ public sealed class DeleteSupplierCommandHandler(
         if(supplier == null)
         {
             return Result.Failure(Error.NotFound($"Supplier with Id {request.Id} not found."));
+        }
+
+        if (supplier.InputReceipts.Any(ir => ir.StatusId == InputStatus.Working))
+        {
+            return Result.Failure(Error.Conflict("Cannot delete supplier with working input receipts."));
         }
 
         deleteRepository.Delete(supplier);
